+++ fpdf/image_parsing.py
@@ -4,7 +4,7 @@ from math import ceil
 from urllib.request import urlopen
 
 try:
-    from PIL import Image
+    from PIL import Image, TiffImagePlugin
 
     try:
         from PIL.Image import Resampling
@@ -65,7 +65,12 @@ def get_img_info(img, image_filter="AUTO", dims=None):
 
     if image_filter == "AUTO":
         # Very simple logic for now:
-        image_filter = "DCTDecode" if img.format == "JPEG" else "FlateDecode"
+        if img.format == "JPEG":
+            image_filter = "DCTDecode"
+        elif img.mode == "1":
+            image_filter = "CCITTFaxDecode"
+        else:
+            image_filter = "FlateDecode"
 
     if img.mode in ("P", "PA") and image_filter != "FlateDecode":
         img = img.convert("RGBA")
@@ -127,6 +132,11 @@ def get_img_info(img, image_filter="AUTO", dims=None):
         ):
             info["smask"] = _to_data(img, image_filter, select_slice=alpha_channel)
 
+    dp = f"/Predictor 15 /Colors {dpn} /Columns {w}"
+
+    if img.mode == "1":
+        dp = f"/BlackIs1 true /Columns {w} /K -1 /Rows {h}"
+
     info.update(
         {
             "w": w,
@@ -134,16 +144,120 @@ def get_img_info(img, image_filter="AUTO", dims=None):
             "cs": colspace,
             "bpc": bpc,
             "f": image_filter,
-            "dp": f"/Predictor 15 /Colors {dpn} /Columns {w}",
+            "dp": dp,
         }
     )
 
     return info
 
 
+class temp_attr:
+    """
+    temporary change the attribute of an object using a context manager
+    """
+
+    def __init__(self, obj, field, value):
+        self.obj = obj
+        self.field = field
+        self.value = value
+
+    def __enter__(self):
+        self.exists = False
+        if hasattr(self.obj, self.field):
+            self.exists = True
+            self.old_value = getattr(self.obj, self.field)
+        setattr(self.obj, self.field, self.value)
+
+    def __exit__(self, exctype, excinst, exctb):
+        if self.exists:
+            setattr(self.obj, self.field, self.old_value)
+        else:
+            delattr(self.obj, self.field)
+
+
+def ccitt_payload_location_from_pil(img):
+    """
+    returns the byte offset and length of the CCITT payload in the original TIFF data
+    """
+    # assert(img.info["compression"] == "group4")
+
+    # Read the TIFF tags to find the offset(s) of the compressed data strips.
+    strip_offsets = img.tag_v2[TiffImagePlugin.STRIPOFFSETS]
+    strip_bytes = img.tag_v2[TiffImagePlugin.STRIPBYTECOUNTS]
+
+    # PIL always seems to create a single strip even for very large TIFFs when
+    # it saves images, so assume we only have to read a single strip.
+    # A test ~10 GPixel image was still encoded as a single strip. Just to be
+    # safe check throw an error if there is more than one offset.
+    if len(strip_offsets) != 1 or len(strip_bytes) != 1:
+        raise NotImplementedError(
+            "Transcoding multiple strips not supported by the PDF format"
+        )
+
+    (offset,), (length,) = strip_offsets, strip_bytes
+
+    return offset, length
+
+
+def transcode_monochrome(img):
+    """
+    Convert the open PIL.Image imgdata to compressed CCITT Group4 data.
+
+    """
+    # Convert the image to Group 4 in memory. If libtiff is not installed and
+    # Pillow is not compiled against it, .save() will raise an exception.
+    newimgio = BytesIO()
+
+    # we create a whole new PIL image or otherwise it might happen with some
+    # input images, that libtiff fails an assert and the whole process is
+    # killed by a SIGABRT:
+    img2 = Image.frombytes(img.mode, img.size, img.tobytes())
+
+    # Since version 8.3.0 Pillow limits strips to 64 KB. Since PDF only
+    # supports single strip CCITT Group4 payloads, we have to coerce it back
+    # into putting everything into a single strip. Thanks to Andrew Murray for
+    # the hack.
+    #
+    # Since version 8.4.0 Pillow allows us to modify the strip size explicitly
+    tmp_strip_size = (img.size[0] + 7) // 8 * img.size[1]
+    if hasattr(TiffImagePlugin, "STRIP_SIZE"):
+        # we are using Pillow 8.4.0 or later
+        with temp_attr(TiffImagePlugin, "STRIP_SIZE", tmp_strip_size):
+            img2.save(newimgio, format="TIFF", compression="group4")
+    else:
+        # only needed for Pillow 8.3.x but works for versions before that as
+        # well
+        pillow__getitem__ = TiffImagePlugin.ImageFileDirectory_v2.__getitem__
+
+        def __getitem__(self, tag):
+            overrides = {
+                TiffImagePlugin.ROWSPERSTRIP: img.size[1],
+                TiffImagePlugin.STRIPBYTECOUNTS: [tmp_strip_size],
+                TiffImagePlugin.STRIPOFFSETS: [0],
+            }
+            return overrides.get(tag, pillow__getitem__(self, tag))
+
+        with temp_attr(
+            TiffImagePlugin.ImageFileDirectory_v2, "__getitem__", __getitem__
+        ):
+            img2.save(newimgio, format="TIFF", compression="group4")
+
+    # Open new image in memory
+    newimgio.seek(0)
+    newimg = Image.open(newimgio)
+
+    offset, length = ccitt_payload_location_from_pil(newimg)
+
+    newimgio.seek(offset)
+    return newimgio.read(length)
+
+
 def _to_data(img, image_filter, **kwargs):
     if image_filter == "FlateDecode":
         return _to_zdata(img, **kwargs)
+
+    if image_filter == "CCITTFaxDecode":
+        return transcode_monochrome(img)
 
     if img.mode == "LA":
         img = img.convert("L")