+++ test/image/image_types/test_insert_images.py
@@ -129,6 +129,13 @@ def test_insert_gif(tmp_path):
     assert_pdf_equal(pdf, HERE / "image_types_insert_gif.pdf", tmp_path)
 
 
+def test_insert_g4_tiff(tmp_path):
+    pdf = fpdf.FPDF()
+    pdf.add_page()
+    pdf.image(HERE / "test.tiff", x=15, y=15)
+    assert_pdf_equal(pdf, HERE / "image_types_insert_tiff.pdf", tmp_path)
+
+
 def test_insert_pillow(tmp_path):
     pdf = fpdf.FPDF()
     pdf.add_page()